apzkr-pzpi-21-7-dronov-illia/Task1-Server/requirements.txt
alembic==1.11.1
asyncpg==0.28.0
email-validator==2.0.0.post2
fastapi==0.100.0
fastapi-pagination==0.12.6
passlib==1.7.4
pre-commit==3.5.0
pydantic-settings==2.0.1
python-decouple==3.8
PyJWT==2.7.0
python-dotenv==0.21.1
redis==4.6.0
regex==2024.4.28
SQLAlchemy==2.0.18
uvicorn==0.22.0
xlsxwriter==3.2.0

